Remove unnecessary overrides from material.min.css
All CSS frameworks come with a lot of things that nobody ever wants. This like overriding the font in our body or styling h6 which is also styled later makes things slow and can give poor ranking in googl page speed insights. It also removes all the polyfills that are not really needed, even with those polyfills in here, other component somewhere that doesn't have the polyfills will surely break.